pyproject.toml
[tool.poetry]
name = "llama32-vision"
version = "0.1.0"
description = "Testing Llama3.2-vision model"
authors = ["hongthana <thanawat.rn@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^10.0.0"
ollama = "^0.3.3"
gradio = "^5.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
